Moved navigator annotations => org.vaadin.spring.navigator.annotation
