+++ samples/mvp-sample/src/main/java/org/vaadin/spring/samples/mvp/ui/view/BannerView.java
@@ -4,7 +4,7 @@ import javax.inject.Inject;
 
 import org.springframework.core.env.Environment;
 import org.vaadin.spring.annotation.VaadinUIScope;
-import org.vaadin.spring.navigator.VaadinView;
+import org.vaadin.spring.navigator.annotation.VaadinView;
 import org.vaadin.spring.samples.mvp.security.config.Scheme;
 import org.vaadin.spring.samples.mvp.ui.component.listener.LogoutLinkListener;
