+++ samples/mvp-sample/src/main/java/org/vaadin/spring/samples/mvp/ui/view/FooterView.java
@@ -1,7 +1,7 @@
 package org.vaadin.spring.samples.mvp.ui.view;
 
 import org.vaadin.spring.annotation.VaadinUIScope;
-import org.vaadin.spring.navigator.VaadinView;
+import org.vaadin.spring.navigator.annotation.VaadinView;
 
 import com.vaadin.navigator.View;
 import com.vaadin.navigator.ViewChangeListener.ViewChangeEvent;
