+++ samples/mvp-sample/src/main/java/org/vaadin/spring/samples/mvp/ui/view/TabPanelView.java
@@ -10,7 +10,7 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.vaadin.spring.annotation.VaadinUIScope;
 import org.vaadin.spring.navigator.SpringViewProvider;
-import org.vaadin.spring.navigator.VaadinView;
+import org.vaadin.spring.navigator.annotation.VaadinView;
 import org.vaadin.spring.samples.mvp.ui.component.grid.DataGrid;
 import org.vaadin.spring.samples.mvp.ui.component.listener.TabSelectedListener;
 import org.vaadin.spring.samples.mvp.ui.component.nav.NavElement;
