+++ samples/sidebar-sample/src/main/java/org/vaadin/spring/samples/sidebar/PlanningView2.java
@@ -22,7 +22,7 @@ import com.vaadin.ui.Label;
 import com.vaadin.ui.VerticalLayout;
 
 import org.vaadin.spring.annotation.VaadinUIScope;
-import org.vaadin.spring.navigator.VaadinView;
+import org.vaadin.spring.navigator.annotation.VaadinView;
 import org.vaadin.spring.sidebar.FontAwesomeIcon;
 import org.vaadin.spring.sidebar.SideBarItem;
