+++ spring-vaadin-sidebar/src/main/java/org/vaadin/spring/sidebar/SideBarUtils.java
@@ -17,11 +17,12 @@ package org.vaadin.spring.sidebar;
 
 import com.vaadin.navigator.View;
 import com.vaadin.ui.UI;
+
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.context.ApplicationContext;
 import org.vaadin.spring.i18n.I18N;
-import org.vaadin.spring.navigator.VaadinView;
+import org.vaadin.spring.navigator.annotation.VaadinView;
 
 import java.util.ArrayList;
 import java.util.Collection;
